add http headers for caching and expiration
This will allow browser and proxy caches to hold on to assets longer, but used in combination with a version string (eg. images/foo.png?1244343425), we can invalidate those caches and force them to fetch a new file when we want. (considered using Jammit-Sinatra, but Steve mentioned this project will likely be updated to Rails, so I'm trying to keep things simple)